\name{draw}
\Rdversion{1.1}
\alias{draw}
\title{
  Render a view of a data set.
}
\description{
  This function is used to render the current view of a data set.
  This is a generic function.
}
\usage{
draw(v)
}
\arguments{
  \item{v}{
    An object of class \code{"Viewer"}.
  }
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{ViewerSimple-class}}
}
\keyword{ dynamic }
