\name{viewerDataFrame}
\Rdversion{1.1}
\alias{viewerDataFrame}
\title{
  Prepare a data frame object for viewing.
}
\description{
  This function derives various important attributes of a
  data frame and stores them so that viewing the data frame
  can occur efficiently.
}
\usage{
viewerDataFrame(data)
}
\arguments{
  \item{data}{ A data frame. }
}
\value{
  An object of class \code{"\linkS4class{ViewerDataFrame}"}.
}
\author{Paul Murrell}
\seealso{
  \code{\linkS4class{ViewerDataFrame}}
  and
  \code{\link{viewerData}}
}
\examples{
}
\keyword{ dynamic }
