\name{viewerDevice}
\Rdversion{1.1}
\alias{viewerDevice}
\title{
  A viewing device
}
\description{
  Creates an object representing the device where
  a data source is being viewed.
}
\usage{
viewerDevice()
}
\value{
  An object of class \code{"\linkS4class{ViewerDeviceDefault}"}.
}
\author{Paul Murrell}
\seealso{
  \code{\linkS4class{ViewerDeviceDefault}}
  and
  \code{\link{simpleViewer}}
}
\examples{
}
\keyword{ dynamic }
