\name{viewerDeviceVp}
\Rdversion{1.1}
\alias{viewerDeviceVp}
\title{
  A viewport-based viewing device
}
\description{
  Creates an object representing the device where
  a data source is being viewed.
}
\usage{
viewerDeviceVp(data)
}
\arguments{
  \item{data}{
    The data source being viewed (an object of class
    \code{\linkS4class{ViewerData}})
  }
}
\value{
  An object of class \code{"\linkS4class{ViewerDeviceViewport}"}.
}
\author{Paul Murrell}
\seealso{
  \code{\linkS4class{ViewerDeviceViewport}},
  \code{\link{viewerData}}
  and
  \code{\link{simpleViewer}}
}
\examples{
}
\keyword{ dynamic }
