\name{viewerState}
\Rdversion{1.1}
\alias{viewerState}
\title{
  Initialize viewer state information.
}
\description{
  Creates an object representing the current view of a data source.
}
\usage{
viewerState()
}
\value{
  An object of class \code{"\linkS4class{ViewerStateSimple}"}.
}
\author{Paul Murrell}
\seealso{
  \code{\linkS4class{ViewerStateSimple}}
  and
  \code{\link{simpleViewer}}
}
\examples{
}
\keyword{ dynamic }
